Essential Functionality complete and book search complete

+++ app/controllers/lists_controller.rb
@@ -1,6 +1,5 @@
 class ListsController < ApplicationController
 before_action :check_for_login, :only => [:new, :create, :edit]
-before_action :check_for_admin, :only => [:destroy]
 
   def index
     @lists = List.all
@@ -31,15 +30,32 @@ before_action :check_for_admin, :only => [:destroy]
 
   def destroy
     list = List.find params[:id]
-    list.destroy
-    redirect_to lists_path
+    if @current_user.admin == true || @current_user.id == list.user_id
+      list.destroy
+      redirect_to lists_path
+    else
+      redirect_to lists_path
+    end
   end
+
 
   def add
     list = List.find params[:list_id]
     book = Book.find params[:book_id]
     list.books << book
     redirect_to list
+  end
+
+  def remove
+    list = List.find params[:list_id]
+    book = Book.find params[:book_id]
+    target = list.books.all
+    if @current_user.admin == true || @current_user.id == list.user_id
+      list.books.delete(book)
+      redirect_to list
+    else
+      redirect_to list
+    end
   end
